requirements.txt
absl-py==2.2.2
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
asttokens==3.0.0
attrdict==2.0.1
attrs==25.3.0
backcall==0.2.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
clarabel==0.9.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
configargparse==1.7
contourpy==1.1.1
crcmod==1.7
cryptography==45.0.6
cvxpy==1.5.2
cycler==0.12.1
dash==3.0.4
decorator==5.2.1
# diff-surfel-rasterization==0.0.1
ecos==2.0.14
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.14.0
flask==3.0.3
fonttools==4.57.0
fsspec==2025.3.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
gym==0.26.2
gym-notices==0.1.0
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==8.12.3
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.14
kiwisolver==1.4.7
lazy-loader==0.4
lpips==0.1.4
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.2
mmcv-lite==2.2.0
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
narwhals==1.37.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
open3d==0.18.0
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
osqp==1.0.4
oss2==2.17.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==25.0.1
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==6.0.1
plyfile==1.0.3
prompt-toolkit==3.0.51
protobuf==3.20.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybullet==3.2.7
pycparser==2.22
pycryptodome==3.23.0
pygments==2.19.1
pyparsing==3.1.4
pyproject-toml==0.0.12
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2023.4
pywavelets==1.4.1
pyyaml==6.0.2
referencing==0.35.1
requests==2.28.2
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.4.2
rpds-py==0.20.1
rsa==4.9.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
scs==3.2.7.post2
setuptools==60.2.0
simple-knn==0.0.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
toml==0.10.2
tomli==2.2.1
# torch==2.1.1+cu118
# torchaudio==2.1.1+cu118
# torchvision==0.16.1+cu118
tqdm==4.65.2
traitlets==5.14.3
trimesh==4.3.2
triton==2.1.0
typing-extensions==4.9.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
werkzeug==3.0.6
widgetsnbextension==4.0.14
yapf==0.43.0
zipp==3.20.2
